@@ -9,12 +9,15 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.10.15",
     "@nextui-org/react": "^1.0.0-beta.10",
+    "@reduxjs/toolkit": "^1.9.1",
     "@types/react-redux": "^7.1.24",
     "axios": "^1.2.1",
     "classnames": "^2.3.2",
-    "next": "13.0.5",
+    "next": "^13.1.1",
     "next-redux-wrapper": "^8.0.0",
     "next-themes": "^0.2.1",
     "npm": "^9.2.0",
