@@ -1,8 +1,9 @@
-// "use client";
-//
-// import { PropsWithChildren } from "react";
-// import { Provider } from "react-redux";
-//
-// export default function ReduxProvider({ children }: PropsWithChildren) {
-//   return <Provider store={store}>{children}</Provider>;
-// }
+"use client";
+
+import { PropsWithChildren } from "react";
+import { Provider } from "react-redux";
+import { store } from "../store";
+
+export default function ReduxProvider({ children }: PropsWithChildren) {
+  return <Provider store={store}>{children}</Provider>;
+}
